// src/components/React Query with TS/src/types/apiData.types.ts
// Defining the Interface based on your API structure
export interface UserData {
  id: string;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo?: {
      lat?: string;
      lng?: string;
    }
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}